chore: delete unused codes

+++ src/main/java/linkfit/service/CareerService.java
@@ -33,17 +33,9 @@ public class CareerService {
     }
 
     public void deleteCareer(Long careerId) {
-        isExistCareer(careerId);
+        if(careerRepository.existsById(careerId)) {
+            throw new NotFoundException(NOT_FOUND_CAREER);
+        }
         careerRepository.deleteById(careerId);
-    }
-
-    public Trainer getTrainerByCareerId(Long careerId) {
-        Career career = isExistCareer(careerId);
-        return career.getTrainer();
-    }
-
-    private Career isExistCareer(Long careerId) {
-        return careerRepository.findById(careerId)
-            .orElseThrow(() -> new NotFoundException(NOT_FOUND_CAREER));
     }
 }
